refactor(astra-connectors): improve separation of concerns and code organization
Split monolithic postgres.rs and postgres_destination.rs into focused modules:

**postgres module structure:**
- mod.rs: PostgresSource orchestrator, spec parsing, SQL utilities
- types.rs: config and result types (PostgresSourceConfig, SnapshotExecutionReport, etc)
- connection.rs: connection management and testing (test_postgres_connection, connect)
- discover.rs: schema introspection (discover_tables)
- snapshot.rs: snapshot execution loop, SQL builders, row encoding

**postgres_destination module structure:**
- mod.rs: PostgresDestinationLoader orchestrator
- types.rs: config and result types (PostgresDestinationConfig, RawLoadReport)
- loader.rs: chunk loading, JSONL decompression, SQL identifier helpers

**Key improvements:**
- Eliminated duplication of require_string/optional_string/require_u16/resolve_password_ref
  handlers via new config_parser.rs shared module
- Single responsibility per module: types are pure data, connection.rs handles I/O,
  discover.rs queries schema, snapshot.rs executes snapshots, loader.rs loads chunks
- PostgresDestinationConfig now embeds PostgresConnectionConfig instead of repeating
  all 7 connection fields
- Tests co-located with their code (snapshot tests in snapshot.rs, etc)
- Public API re-exports remain unchanged in lib.rs — this is a pure internal refactor

All 17 unit tests pass. CLI and control-plane integration tests also pass with Podman.
